Add a new filed client contract address for allocator
