+++ manual-migrations/2024-10-16.sql
@@ -1,0 +1,2 @@
+ALTER TABLE IF EXISTS public.allocators
+    ADD COLUMN client_contract_address text;
